web/requirements.txt
asgiref==3.2.10
Django==3.1.1
django-cors-headers==3.5.0
djangorestframework==3.11.1
psycopg2==2.9.2
pytz==2021.3
sqlparse==0.4.2
